pyproject.toml
[tool.poetry]
name = "flashmonsters-api"
version = "0.1.0"
description = ""
authors = ["ShinnosukeUesaka <shinnosuke.uesaka@gmail.com>"]
readme = "README.md"
packages = [{include = "flashmonsters_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
djangorestframework = "^3.14.0"
openai = "^0.28.1"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
